LayoutTests/imported/w3c/web-platform-tests/navigation-api/currententrychange-event/navigate-from-initial-about-blank-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL
CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL


PASS currententrychange does not fire when navigating away from the initial about:blank
